Move segment-based error budget analysis into modal_analysis
